+++ assets/shaders/main_unlit.vs.glsl
@@ -37,7 +37,7 @@ layout (std140) uniform BatchDataBlock {
     ivec4 uInstanceIndices[$MaxBatchSize / 4];
 };
 
-layout(binding = $InstanceBufSlot) uniform mediump samplerBuffer instances_buffer;
+layout(binding = $InstanceBufSlot) uniform highp samplerBuffer instances_buffer;
 
 #ifdef VULKAN
 layout(location = 4) out vec2 aVertexUVs1_;
